Move FaissHookFvec/distances_simd/BitsetView into knowhere/utils
