+++ knowhere/knowhere/index/vector_index/VecIndex.h
@@ -11,7 +11,6 @@
 
 #pragma once
 
-#include <faiss/utils/BitsetView.h>
 #include <memory>
 #include <utility>
 #include <vector>
@@ -24,6 +23,7 @@
 #include "knowhere/index/IndexType.h"
 #include "knowhere/index/vector_index/Statistics.h"
 #include "knowhere/index/vector_index/helpers/DynamicResultSet.h"
+#include <knowhere/utils/BitsetView.h>
 
 namespace milvus {
 namespace knowhere {
