+++ knowhere/knowhere/utils/distances_simd.h
@@ -11,9 +11,6 @@
 
 #include <stdint.h>
 
-#include <faiss/utils/BitsetView.h>
-
-
 namespace faiss {
 
  /*********************************************************
