+++ knowhere/knowhere/utils/distances_simd_avx.cpp
@@ -1,12 +1,12 @@
 
 // -*- c++ -*-
 
-#include <faiss/utils/distances_simd_avx.h>
-
 #include <cstdio>
 #include <cassert>
 #include <immintrin.h>
 #include <string>
+
+#include "distances_simd_avx.h"
 
 namespace faiss {
