+++ knowhere/knowhere/utils/distances_simd_avx.h
