+++ unittest/test_common.cpp
@@ -12,9 +12,9 @@
 #include <gtest/gtest.h>
 #include "knowhere/common/Dataset.h"
 #include "knowhere/common/Timer.h"
-#include "knowhere/knowhere/common/Exception.h"
+#include "knowhere/common/Exception.h"
+#include "knowhere/utils/BitsetView.h"
 #include "unittest/utils.h"
-#include "faiss/utils/BitsetView.h"
 #include <boost/dynamic_bitset.hpp>
 
 /*Some unittest for knowhere/common, mainly for improve code coverage.*/
